# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DenTeeth API Documentation",
        "description": "API Documentation for DenTeeth",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Development server"
        },
        {
            "url": "https://www.example.com",
            "description": "Production server"
        }
    ],
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get all addresses",
                "description": "Returns all addresses",
                "operationId": "33ccea30443418dbab9242aa67931614",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Create a new address",
                "description": "Create a new address",
                "operationId": "5549e34fec9817a9794302a767fc7fde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "house_number",
                                    "street_name",
                                    "street_number",
                                    "commune",
                                    "district",
                                    "province",
                                    "city",
                                    "postal_code"
                                ],
                                "properties": {
                                    "house_number": {
                                        "type": "integer",
                                        "example": "123"
                                    },
                                    "street_name": {
                                        "type": "string",
                                        "example": "Main Street"
                                    },
                                    "street_number": {
                                        "type": "integer",
                                        "example": "456"
                                    },
                                    "commune": {
                                        "type": "string",
                                        "example": "Commune"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "District"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Province"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "City"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address added successfully"
                    },
                    "422": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/addresses/{id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get a specific address",
                "description": "Returns a specific address by ID",
                "operationId": "de6419a46a2f3639437add4f3cb76e5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Update an existing address",
                "description": "Update an existing address",
                "operationId": "2f078db19c8d7e5a374faa4b048e4c45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "house_number": {
                                        "type": "integer",
                                        "example": "123"
                                    },
                                    "street_name": {
                                        "type": "string",
                                        "example": "Main Street"
                                    },
                                    "street_number": {
                                        "type": "integer",
                                        "example": "456"
                                    },
                                    "commune": {
                                        "type": "string",
                                        "example": "Commune"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "District"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Province"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "City"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address updated successfully"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete an address",
                "description": "Delete an address by ID",
                "operationId": "3380f49103f9fb0f4a126d324894a33a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted successfully"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}